Try to use non-conflicting, platform-dependent signals instead of SIGQUIT.  (Darwin uses SIGEMT, which doesn't seem to be raised by the kernel on x86 or PPC; other platfroms can use user-defined (possibly "realtime" signals).
Rename kernel things that had 'quit' in their names and had to
do with terminating threads to instead have 'thread_kill' in
their names (e.g. quit_handler -> thread_kill_handler.)

In the x86 GC, at least in mark_root and the recursive case of
rmark, check the dnode against gc_area_dnode before mapping
a TRA to the containing function.  (This keeps us from crashing
in those cases if we see a garbage root that's tagged as a TRA,
but that fixes the symptom and not the proble, that would cause
such a garbage root to appear.)  This is x86-specific; the
PPC ports don't use TRAs.

Save lisp_heap_gc_threshold, the EGC enable state, and the
sizes of the ephemeral generations in the image and restore
them on startup.  (The -T option - which sets the global
GC threshold from the command line - overrides any value
set in the image; more accurately, it does that if the value
of the -T argument isn't the default GC threshold size.)
This is probably the right idea, but it's an incompatible
change; people who override the defaults at build or startup
time may need to change how their systems are built and
initialized.