Try to address ticket:649 by clearly distinguishing between UnMapMemory() (which completely deallocates an entire memory region created by MapMemory() or MapMemoryForStack()) from UnCommitMemory() (which leaves a previously mapped and accessible range of pages mapped but inaccessible/uncommitted.)
Follow Alexander Gavrilov's suggestions (and use his patch) to ensure
that thread handles are closed on Windows.  Make create_system_thread()
return a Boolean, since no callers care about the exact value and the
value that was returned on Windows was a handle that needed to be closed.

Ensure that create_system_thread() observes its stack_size argument.
Make the size of vstack soft protected area ("the yellow zone") larger.

Use a readable/writable PAGE_GUARD page at the bottom of the
cstack/tstack yellow zone and handle the Windows exception that's
raised when a PAGE_GUARD page is written to, so that stack overflow
detection has a prayer of working on Windows.

UNTESTED ANYWHERE BUT WIN32.