In _SPcallback, save the caller's mxcsr on and restore it from the stack, not in tcr.foreign_mxcsr.  Different callers can legitimately have different mxcsr values (as long as they save/restore the mxcsr before changing its value), and we want callbacks to restore the caller's value (not that of the most recent callback on the thread.)
callback_to_lisp() in the CCL kernel is an example of a C function that
calls into lisp with an atypical value in the mxcsr.  It's not clear
that that's a useful value, but callback_to_lisp() saves/restores the
mxcsr before setting it to that value. 

Reported problems which are symptoms of the mxcsr getting trashed in
the IDE seem to stem from the fact that callback_to_lisp() invokes a
callback with this atypical mxcsr value when unresolved EEPs are used
from callbacks.  We might want to change callback_to_lisp() to not do
that, but the problem is and always has been that callbacks don't
restore the mxcsr that they were called with.  The treatment of 
the mxcsr by _SPffcall (and the behavior of callback_to_lisp()) just
exposed that problem.