Change the rules a bit to allow the LR to point to arbitrary word-aligned locations inside a vector of single- or double-floats. The old (well, few months old) trick of allowing the LR to point to the (0-valued) pad word at the beginning of a DOUBLE-FLOAT or DOUBLE-FLOAT vector isn't supported anymore (if we can point the lr at any word in the vector, we can't tell whether that's the pad word or not.) This is GC safe as long as the containing vector is marked before the LR is (and that's true of the ARM version of mark_xp()). Having the LR point into an ivector (other than a CODE-VECTOR) isn't sufficient to mark the ivector, so we generally have to be careful to ensure that the LR doesn't point into an ivector that becomes garbage.  Vinsns that use the LR as a locative zero it out after use; that's correct but overly conservative.  (We -could- have the compiler watch the vinsn stream and only zero the LR if the register containing the LR changes/might change and cause the vector to get GCed  before the LR itself changes.  We might want that sort of vinsn-watching mechanism for other reasons, but zeroing the LR isn't incredibly expensive and it's much better to be safe than sorry.)
The point(s) of all this is/are to avoid the mandatory use of
imm0/imm1 in fp vector references and to make it practical to (at
least in LAP) use NEON vector instructions in some cases.  (Because of
addressing limitations in the ARM's coprocessor load/store
instructions, we pretty much have to have a way to point "into"
floating-point vectors; we won't have the option of loading/storing
GPR pairs when using quadword NEON instructions.)

This is an ABI change, though I don't plan to change image/fasl
versions in the trunk.  To be safe, recompile any FP-intensive 
ARM code after doing a full rebuild.