nx2.lisp: New function NX2-VAR-NOT-REFFED-BY-FORM-P.  Walks acode, returns true iff it's sure that acode doesn't reference the specifed VAR.
x862.lisp: in the fixed-arg tail-call case of X862-SELF-CALL, don't
process arguments whose value isn't changing between the caller and 
callee.  If argument values are ultimately going to wind up in NVRs,
try to get them there directly.  (This is only safe if no subequently-
processed argument values reference or set the variable's value.  Setting
the NVR changes that value, and N-TARGETED-REG-FORMs isn't used to dealing
with that side-effect.)