@+ vinsn macro doesn't invalidate regmap.  (Spent several days trying to avoid regmap invalidation in some cases involving nested IF/AND, but didn't commit that.)
When loading from the vstack into a non-node target, try to use
a non-conflicting temporary.

When eliding-pushes on the vstack, some cases depend on the order
in which things happen in the sequence.  We can't really be sure
of that order unless the sequence is straight-line code.

Fix typos in ARM2-FOUR-UNTARGETED-REG-FORMS.

If a variable's in an FPR, then things like (SYMBOLP VAR) may try
to compare the FPR to NIL.  Compare the SP to NIL instead (we're
just trying to clear the Z bit in the status register.)

ARM2-%SINGLE-FLOAT-NEGATE should use ! single-float-negate, not the
double-float variant.

If the result of %NATURAL-LOGAND is known to be small enough and the
result needs to be boxed, boxing is simpler.