In the definition of current_stack_pointer(), use different inline asm for the x86 targets.
This works around a bug in the llvm-gcc that Apple
ships with Xcode 4.2.1 (and maybe earlier versions).
When compiling with -g without -O, the previous inline
asm is miscompiled.

On my 10.7.2 system with Xcode 4.2.1, /usr/bin/cc
is a link to llvm-gcc.  cc --verison reports:
i686-apple-darwin11-llvm-gcc-4.2 (GCC) 4.2.1 (Based
on Apple Inc. build 5658) (LLVM build 2336.1.00)