lib/compile-ccl.lisp: Define some "optional features" for testing hashing algorithms.  :EQ-HASH-MONITOR causes EQ-hashing routines to record the number of calls/probes involved.  (The INCFs here aren't thread-safe.)  :MIXUP-HASH-CODE-NOP makes the function CCL::MIXUP-HASH-CODE return its argument.
Implement WITH-GLOBAL-OPTIMIZATION-SETTINGS a little differently;
use it in TEST-CCL.

level-0/X86/x86-hash.lisp: in STRIP-TAG-TO-FIXNUM, shift by an
extra bit (so that results aren't always odd/even depending on
tag bits of arg.)  This seems to be the right thing and should
likely be implemented on all architectures.

level-0/l0-hash.lisp: conditionally implement the new optional
features.

NEED-USE-EQL: can use EQ if arg is a fixnum or #+64-bit-target
a SINGLE-FLOAT.