Conditionally load the IDE init file in the delegate's #/applicationDidFinishLaunching: method instead of in its #/applicationWillFinishLaunching: method (which is called earlier); the global NSApplication object may not be usefully initialized when the earlier method is called.
Continue to call INIT-USER-INTERFACE - which doesn't depend on the
state of the NSApplication object - from
#/applicationWillFinishLaunching:, to ensure that the global
NSDocumentController object is an instance of our subclass.

Don't bother to call (or define) MAYBE-FIXUP-APPLICATION-MENU at all:
this was originally intended to replace instances of the IDE bundle name
(e.g., "Clozure CL64") with something more generic ("Clozure CL"), but it
never worked (did string comparisons with the wrong args and was generally
broken) and people have likely gotten accustomed to the current behavior.