Fix the loop facility so that (loop named nil ...) actually generates a block named nil.
A CL loop form can have only one named clause, so in loop-do-named,
simply set *loop-names* to a list containing the symbol that should be
the loop's name.

In loop-translate, we know that *loop-names* will be either nil or a
list of one symbol, so just use (car *loop-names*) as the name of the
loop's implicit block.

Closes ticket:1099.