Use the exception interface provided in mach/mach_exc.defs.
Unlike exc_server/catch_exception_raise/etc., which are in the system
library, to use the mach_exc_server/catch_mach_exception_raise/etc.
interface, we must run mig to generate the source files from the
interface definiton.  So, update the Darwin makefiles to do that.

I didn't see any reason to keep using the -exported_symbols_list
linker option, so I removed that, too.

The main thing this gets us is 64-bit exception codes.  (But we don't
take advantage of this yet.)