New file ccl:mac-ui;ccl-application.lisp.
The IDE now defines ide-application, which inherits from ccl-application.
This replaces the class lisp-application, which is a too generic name.
Rename the delegate class name from lisp-application-delegate to
ide-application-delegate based on a similar rationale.

The function start-cocoa-ide now takes zero arguments, and uses the
special variable *delegate-class-name* to determine what class to
instantiate to use for the application delegate.

Update Info.plist-proto to use IDEApplication as the application class.

Define IDE and CCL as special objc words in defsystem.lisp.