Stop using a repz prefix on certain ret instructions
Adding the prefix apparently helped branch prediction on some
AMD hardware.  Starting with AMD family 15h (Bulldozer), which
came out in 2011, this is no longer needed or recommended.

This is probably not very significant, but it removes a little
complexity, which seems like it can only be a good thing.