Update compiler-macro on make-array to recognize '0
In CCL, newly allocated arrays are zero-filled, so we try to
recognize :initial-element cases where we can skip pointless
re-zeroing of the array.

We already recognized :initial-element 0, but there is code in
the wild that does :initial-element '0, so we want to recognize
that as well.

Make a few stylistic tweaks to that compiler macro as well.